// frontend/types/user.ts
export interface UserInfo {
  name: string;
  email: string;
  phone: string;
  designation: string;
  organization: string;
  address: string;
}

export interface UserRow {
  id: string;
  email: string;
  full_name: string;
  phone: string;
  designation: string;
  organization: string;
  address: string;
  updated_at: string;
}